pyproject.toml
[tool.poetry]
name = "hashchat"
version = "0.1.0"
description = "HashTag based chat"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.6"
channels = "^3.0.3"
channels-redis = "^3.2.0"
django-allauth = "^0.44.0"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
